# Conflicts:
#	security-admin-sql/pom.xml

+++ security-admin-impl/src/main/java/net/n2oapp/security/admin/impl/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package net.n2oapp.security.admin.impl.service;
 
+import net.n2oapp.platform.i18n.UserException;
 import net.n2oapp.security.admin.api.criteria.UserCriteria;
 import net.n2oapp.security.admin.api.model.*;
 import net.n2oapp.security.admin.api.provider.SsoUserRoleProvider;
@@ -13,13 +14,16 @@ import net.n2oapp.security.admin.impl.repository.RoleRepository;
 import net.n2oapp.security.admin.impl.repository.UserRepository;
 import net.n2oapp.security.admin.impl.service.specification.UserSpecifications;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collections;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static java.util.Objects.isNull;
@@ -44,6 +48,12 @@ public class UserServiceImpl implements UserService {
     private UserValidations userValidations;
     @Autowired
     private AuditHelper audit;
+
+    @Value("${access.user.send-mail-delete-user:false}")
+    private Boolean sendMailDelete;
+
+    @Value("${access.user.send-mail-activate-user:false}")
+    private Boolean sendMailActivate;
 
     public UserServiceImpl(UserRepository userRepository, RoleRepository roleRepository, SsoUserRoleProvider provider) {
         this.userRepository = userRepository;
@@ -73,7 +83,7 @@ public class UserServiceImpl implements UserService {
         UserEntity savedUser = userRepository.save(entity);
         // получаем модель для сохранения SSO провайдером, ему надо передать незакодированный пароль
         if (provider.isSupports(savedUser.getExtSys())) {
-            SsoUser ssoUser = ssoModel(savedUser);
+            SsoUser ssoUser = model(savedUser);
             ssoUser.setPassword(password);
             // если пароль временный, то требуем смены пароля при следующем входе пользователя
             if (user.getTemporaryPassword() != null)
@@ -103,6 +113,10 @@ public class UserServiceImpl implements UserService {
             userValidations.checkPassword(user.getPassword(), user.getPasswordCheck(), user.getId());
         }
         UserEntity entityUser = userRepository.getOne(user.getId());
+        boolean isActiveChanged = !Objects.equals(entityUser.getIsActive(), user.getIsActive());
+        if (entityUser.getUsername().equals(getContextUserName()) && isActiveChanged) {
+            throw new UserException("exception.selfChangeActivity");
+        }
         entityUser = entityForm(entityUser, user);
         // кодируем пароль перед сохранением в бд если он изменился
         if (nonNull(user.getPassword()))
@@ -110,7 +124,7 @@ public class UserServiceImpl implements UserService {
         UserEntity updatedUser = userRepository.save(entityUser);
         //в провайдер отправляем незакодированный пароль, если он изменился, и отправляем null, если не изменялся пароль
         if (provider.isSupports(updatedUser.getExtSys())) {
-            SsoUser ssoUser = ssoModel(updatedUser);
+            SsoUser ssoUser = model(updatedUser);
             if (isNull(user.getPassword())) {
                 ssoUser.setPassword(null);
             } else {
@@ -118,14 +132,24 @@ public class UserServiceImpl implements UserService {
             }
             provider.updateUser(ssoUser);
         }
-        return audit("audit.userUpdate", model(updatedUser));
+        User result = model(updatedUser);
+        if (sendMailActivate && isActiveChanged) {
+            mailService.sendChangeActivateMail(result);
+        }
+        return audit("audit.userUpdate", result);
     }
 
     @Override
     public void delete(Integer id) {
-        SsoUser user = ssoModel(userRepository.findById(id).orElse(null));
+        SsoUser user = model(userRepository.findById(id).orElse(null));
+        if (nonNull(user) && user.getUsername().equals(getContextUserName())) {
+            throw new UserException("exception.selfDelete");
+        }
         userRepository.deleteById(id);
         if (nonNull(user)) {
+            if (sendMailDelete) {
+                mailService.sendUserDeletedMail(user);
+            }
             audit("audit.userDelete", user);
             if (provider.isSupports(user.getExtSys())) provider.deleteUser(user);
         }
@@ -156,10 +180,16 @@ public class UserServiceImpl implements UserService {
     @Override
     public User changeActive(Integer id) {
         UserEntity userEntity = userRepository.findById(id).orElse(null);
+        if (nonNull(userEntity) && userEntity.getUsername().equals(getContextUserName())) {
+            throw new UserException("exception.selfChangeActivity");
+        }
         userEntity.setIsActive(!userEntity.getIsActive());
-        SsoUser result = ssoModel(userRepository.save(userEntity));
+        SsoUser result = model(userRepository.save(userEntity));
         if (provider.isSupports(userEntity.getExtSys())) {
             provider.changeActivity(result);
+        }
+        if (sendMailActivate) {
+            mailService.sendChangeActivateMail(model(userEntity));
         }
         return audit("audit.userChangeActive", result);
     }
@@ -210,7 +240,7 @@ public class UserServiceImpl implements UserService {
 
                 //в провайдер отправляем незакодированный пароль
                 if (provider.isSupports(updatedUser.getExtSys())) {
-                    SsoUser ssoUser = ssoModel(updatedUser);
+                    SsoUser ssoUser = model(updatedUser);
                     ssoUser.setPassword(password);
                     // если пароль временный, то требуем смены пароля при следующем входе пользователя
                     if (user.getTemporaryPassword() != null)
@@ -226,6 +256,7 @@ public class UserServiceImpl implements UserService {
     }
 
     private UserEntity entityForm(UserEntity entity, UserForm model) {
+        entity.setIsActive(Boolean.TRUE.equals(entity.getIsActive()));
         entity.setUsername(model.getUsername());
         entity.setName(model.getName());
         entity.setSurname(model.getSurname());
@@ -241,6 +272,7 @@ public class UserServiceImpl implements UserService {
             entity.setRoleList(model.getRoles().stream().filter(roleId -> roleId > 0).map(RoleEntity::new).collect(Collectors.toList()));
         return entity;
     }
+
 
     private UserEntity entityProvider(SsoUser modelFromProvider) {
         UserEntity entity = new UserEntity();
@@ -266,7 +298,20 @@ public class UserServiceImpl implements UserService {
         return entity;
     }
 
-    private User model(UserEntity entity) {
+    private String getContextUserName() {
+        if (nonNull(SecurityContextHolder.getContext()) &&
+                nonNull(SecurityContextHolder.getContext().getAuthentication()) &&
+                SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+                        instanceof org.springframework.security.core.userdetails.User) {
+            return ((org.springframework.security.core.userdetails.User)
+                    SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
+        }
+
+        return null;
+    }
+
+
+    private SsoUser model(UserEntity entity) {
         if (isNull(entity)) return null;
         SsoUser model = new SsoUser();
         model.setId(entity.getId());
@@ -282,6 +327,8 @@ public class UserServiceImpl implements UserService {
         model.setDepartment(model(entity.getDepartment()));
         model.setOrganization(model(entity.getOrganization()));
         model.setRegion(model(entity.getRegion()));
+        model.setExtSys(entity.getExtSys());
+        model.setExtUid(entity.getExtUid());
         StringBuilder builder = new StringBuilder();
         if (nonNull(entity.getSurname())) {
             builder.append(entity.getSurname()).append(" ");
@@ -300,13 +347,6 @@ public class UserServiceImpl implements UserService {
             }).collect(Collectors.toList()));
         }
         return model;
-    }
-
-    private SsoUser ssoModel(UserEntity entity) {
-        SsoUser ssoUser = (SsoUser) model(entity);
-        ssoUser.setExtSys(entity.getExtSys());
-        ssoUser.setExtUid(entity.getExtUid());
-        return ssoUser;
     }
 
     private Role model(RoleEntity entity) {
